feat: refactor user type imports to use the correct User interface
