+++ src/interfaces/User.ts
